Move homepage testing code to proper place
